refactored algorithms to avoid empty strings
