+++ src/main/java/pl/jedro/recognitionApp/exceptions/NameIsEmptyException.java
@@ -1,0 +1,4 @@
+package pl.jedro.recognitionApp.exceptions;
+
+public class NameIsEmptyException extends IllegalArgumentException {
+}
