--- src/main/java/pl/jedro/recognitionApp/exceptions/PathNotSpecifiedException.java
@@ -1,14 +1,0 @@
-package pl.jedro.recognitionApp.exceptions;
-
-import java.io.FileNotFoundException;
-
-public class PathNotSpecifiedException extends FileNotFoundException {
-
-
-    public PathNotSpecifiedException() {
-    }
-
-    public PathNotSpecifiedException(String s) {
-        super(s);
-    }
-}
