+++ src/main/java/pl/jedro/recognitionApp/models/Genders.java
@@ -1,4 +1,4 @@
-package pl.jedro.recognitionApp.model;
+package pl.jedro.recognitionApp.models;
 
 public enum Genders {
     MALE, FEMALE, INCONCLUSIVE,
