+++ src/main/java/pl/jedro/recognitionApp/utils/GenderTokensBufferedReader.java
@@ -2,8 +2,7 @@ package pl.jedro.recognitionApp.utils;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import pl.jedro.recognitionApp.exceptions.PathNotSpecifiedException;
-import pl.jedro.recognitionApp.model.GenderToken;
+import pl.jedro.recognitionApp.models.GenderToken;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
@@ -17,24 +16,18 @@ public class GenderTokensBufferedReader implements GenderTokensReader {
 
     @Value("${females.path}")
     private String femaleTokensPath;
-
-
     @Override
     public Stream<GenderToken> getMaleTokensStream() throws FileNotFoundException {
         return getGenderTokenStream(maleTokensPath);
     }
-
     @Override
     public Stream<GenderToken> getFemaleTokensStream() throws FileNotFoundException {
         return getGenderTokenStream(femaleTokensPath);
     }
 
     private Stream<GenderToken> getGenderTokenStream(String tokensPath) throws FileNotFoundException {
-        if (maleTokensPath == null) {
-            throw new PathNotSpecifiedException();
-        }
         BufferedReader reader = new BufferedReader(new FileReader(tokensPath));
-        return reader.lines().map(this::mapper);
+        return reader.lines().filter(line -> !line.equals("")).map(this::mapper);
     }
 
     private GenderToken mapper(String s) {
