+++ src/test/java/pl/jedro/recognitionApp/services/GenderRecognitionServiceTests.java
@@ -5,7 +5,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import pl.jedro.recognitionApp.model.Genders;
+import pl.jedro.recognitionApp.models.Genders;
 import pl.jedro.recognitionApp.strategies.AlgorithmFactory;
 import pl.jedro.recognitionApp.strategies.AlgorithmNames;
 import pl.jedro.recognitionApp.utils.GenderTokensReader;
