+++ src/test/java/resources/static/maleTokens.txt
@@ -1,3 +1,5 @@
 Jan
+
 Zbigniew
+
 Karol
